[FIX] Added SettingsButton component
