+++ skill-db/frontend/src/components/EditComponent.vue
@@ -1,54 +1,21 @@
 <template>
-  <div class="d-flex justify-end">
-    <v-menu :close-on-content-click="false">
-      <template v-slot:activator="{ props }">
-        <v-btn v-bind="props"
-               min-width="40px" width="40px" height="35px"
-               class="pa-0" elevation="0">
-          <v-icon size="large">
-            mdi-cog
-          </v-icon>
-        </v-btn>
-      </template>
-      <v-list>
-        <v-list-item link @click.stop="toggleDelete">
-          <v-list-item-title> Löschen</v-list-item-title>
-        </v-list-item>
-        <v-list-item link @click="lockProfile">
-          <v-list-item-title> Sperren</v-list-item-title>
-        </v-list-item>
-      </v-list>
-    </v-menu>
-
-    <v-dialog v-model="toDelete" max-width="200">
-      <v-card>
-        <v-card-text>Willst du dieses Profil wirklich löschen?</v-card-text>
-        <v-card-actions>
-          <v-spacer></v-spacer>
-          <v-btn @click="deleteProfile">
-            Ja
-          </v-btn>
-          <v-btn @click="toggleDelete" class="ml-2">
-            Nein
-          </v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-dialog>
-    <v-overlay v-model="locked" absolute/>
-  </div>
+  <SettingsButton v-if="update === 'true'" :id="this.id" :edit="true"/>
 
   <v-container>
     <v-form @submit.prevent>
       <div class="header">
         <div class="d-flex flex-column align-items-center">
           <img src="@/assets/images/dummy_profilePicture.png" alt="Profilbild">
-          <v-card class="mt-3 mb-10 d-flex justify-center" color="grey" width="200px" height="25px">Bild hochladen</v-card>
+          <v-card class="mt-3 mb-10 d-flex justify-center" color="grey" width="200px" height="25px">Bild hochladen
+          </v-card>
         </div>
         <v-row class="headline">
           <v-col cols="12" lg="6" md="6" sm="12">
             <v-text-field v-model="firstName" :rules="[v => v.length > 1 || 'Erforderlich!']" label="Vorname"/>
-            <v-autocomplete v-model="jobTitle" :items="jobs" label="Jobprofil" :rules="[v => v.length > 1 || 'Erforderlich!']"></v-autocomplete>
-            <v-text-field v-model="phoneNumber" label="Telefonnummer" :rules="[ number => checkPhoneFormat(number) || 'Min. 11 max. 13 Ziffern']"/>
+            <v-autocomplete v-model="jobTitle" :items="jobs" label="Jobprofil"
+                            :rules="[v => v.length > 1 || 'Erforderlich!']"></v-autocomplete>
+            <v-text-field v-model="phoneNumber" label="Telefonnummer"
+                          :rules="[ number => checkPhoneFormat(number) || 'Min. 11 max. 13 Ziffern']"/>
           </v-col>
           <v-col lg="6" md="6" sm="12">
             <v-text-field v-model="lastName" :rules="[v => v.length > 1 || 'Erforderlich']" label="Nachname"/>
@@ -58,7 +25,8 @@
                 label="Primärkompetenz"
                 :rules="[v => v.length > 1 || 'Erforderlich!']"
             />
-            <v-text-field v-model="birthdate" label="Geburtsdatum" :rules="[ date => checkDateFormat(date) || 'Date Format must be DD.MM.YYYY']"/>
+            <v-text-field v-model="birthdate" label="Geburtsdatum"
+                          :rules="[ date => checkDateFormat(date) || 'Date Format must be DD.MM.YYYY']"/>
           </v-col>
         </v-row>
       </div>
@@ -100,47 +68,24 @@
     </v-form>
   </v-container>
 </template>
-<script>
+
+<script lang="js"> /*TODO should be TypeScript*/
 import {ConvertToDetailModelForOutput} from "@/models/DetailModel";
 import router from "@/router";
 import {useDetailStore} from "@/stores/DetailStore";
-import {ref} from "vue";
 import {useRoute} from "vue-router";
+import SettingsButton from "@/components/SettingsButton.vue";
 
 export default {
   props: ['detail', 'update'],
+  components: { SettingsButton },
   setup() {
     const detailStore = useDetailStore();
     const id = Number(useRoute().params.id);
     detailStore.loadDetailsById(id);
-
-    const locked = ref(false);
-    const toDelete = ref(false);
-
-    function enterEdit() {
-      router.push({ name: 'editView', params: { id: this.id}});
-    }
-
-    function toggleDelete() {
-      toDelete.value = !toDelete.value;
-    }
-
-    function lockProfile() {
-      locked.value = true;
-      console.log("This profile is now locked away.");
-      /*router.push(`/`);*/
-    }
-
-    function deleteProfile() {
-      detailStore.deleteDetailsByID(this.id);
-      router.push(`/`);
-    }
-
     return {
-      toDelete, locked,
-      enterEdit, toggleDelete,
-      lockProfile, deleteProfile
-    };
+      id
+    }
   },
   data() {
     return {
@@ -190,7 +135,7 @@ export default {
         const editDetails = ConvertToDetailModelForOutput.toDetail(this);
         const id = this.detail.getId();
         detailStore.updateProfile(editDetails, id);
-        router.push( { name: 'userDetails', params: {id}});
+        router.push({name: 'userDetails', params: {id}});
       } else {
         const newDetails = ConvertToDetailModelForOutput.toDetail(this);
         detailStore.createProfile(newDetails);
@@ -200,7 +145,7 @@ export default {
     leave() {
       if (this.update === 'true') {
         const id = this.detail.getId();
-        router.push( { name: 'userDetails', params: {id}});
+        router.push({name: 'userDetails', params: {id}});
       } else {
         router.push('/');
       }
@@ -232,13 +177,16 @@ export default {
   width: 100%;
   display: flex;
 }
+
 .headline {
   margin-left: 20px;
 }
+
 img {
   height: 200px;
   width: 200px;
 }
+
 .skillsAndDegree {
   margin-left: -10px;
 }
@@ -248,34 +196,42 @@ img {
     display: grid;
     grid-template-rows: 1fr 1fr;
   }
+
   img {
     grid-row: 1;
     margin-left: 0;
   }
+
   .headline {
     margin-left: -10px;
   }
+
   .skillsAndDegree {
     margin-left: 0;
   }
 }
+
 @media screen and (max-width: 957px) {
   .header {
     display: grid;
     grid-template-rows: 0.5fr 1fr;
   }
+
   img {
     grid-row: 1;
     margin-left: 0;
   }
+
   .headline {
     grid-row: 2;
     margin-left: -10px;
     margin-bottom: 10px;
   }
+
   .skillsAndDegree {
     margin-left: 0;
   }
+
   .buttons {
     flex-direction: column;
