+++ skill-db/frontend/src/views/DetailView.vue
@@ -1,45 +1,6 @@
 <template>
   <div v-if="!detailStore.loading">
-    <div class="d-flex justify-end">
-      <v-menu :close-on-content-click="false">
-        <template v-slot:activator="{ props }">
-          <v-btn v-bind="props"
-                 min-width="40px" width="40px" height="35px"
-                 class="pa-0" elevation="0">
-            <v-icon size="large">
-              mdi-cog
-            </v-icon>
-          </v-btn>
-        </template>
-        <v-list>
-          <v-list-item link @click="enterEdit">
-            <v-list-item-title> Bearbeiten</v-list-item-title>
-          </v-list-item>
-          <v-list-item link @click.stop="toggleDelete">
-            <v-list-item-title> Löschen</v-list-item-title>
-          </v-list-item>
-          <v-list-item link @click="lockProfile">
-            <v-list-item-title> Sperren</v-list-item-title>
-          </v-list-item>
-        </v-list>
-      </v-menu>
-
-      <v-dialog v-model="toDelete" max-width="200">
-        <v-card>
-          <v-card-text>Willst du dieses Profil wirklich löschen?</v-card-text>
-          <v-card-actions>
-            <v-spacer></v-spacer>
-            <v-btn @click="deleteProfile">
-              Ja
-            </v-btn>
-            <v-btn @click="toggleDelete" class="ml-2">
-              Nein
-            </v-btn>
-          </v-card-actions>
-        </v-card>
-      </v-dialog>
-      <v-overlay v-model="locked" absolute/>
-    </div>
+    <SettingsButton :id="id" :edit="false"/>
 
     <div class="header">
       <img src="@/assets/images/dummy_profilePicture.png" alt="Profilbild">
@@ -109,44 +70,18 @@
 
 <script lang="ts">
 import {useDetailStore} from "@/stores/DetailStore";
-import {ref} from "vue";
-import router from "@/router";
 import {useRoute} from "vue-router";
+import SettingsButton from "@/components/SettingsButton.vue";
 
 export default {
+  components: { SettingsButton },
   setup() {
     const detailStore = useDetailStore();
     const id = Number(useRoute().params.id);
     detailStore.loadDetailsById(id);
-    const detail = detailStore.details;
-
-    const locked = ref(false);
-    const toDelete = ref(false);
-
-    function enterEdit(): void {
-      router.push({name: 'editView', params: {id: id}});
-    }
-
-    function toggleDelete(): void {
-      toDelete.value = !toDelete.value;
-    }
-
-    function lockProfile(): void {
-      locked.value = true;
-      console.log("This profile is now locked away.");
-      /*router.push(`/`);*/
-    }
-
-    function deleteProfile(): void {
-      detailStore.deleteDetailsByID(id);
-      router.push(`/`);
-    }
 
     return {
-      detail, detailStore,
-      toDelete, locked,
-      enterEdit, toggleDelete,
-      lockProfile, deleteProfile
+      detailStore, id
     };
   },
 }
@@ -166,10 +101,6 @@ img {
 .headline {
   height: 100px;
   margin-left: 10%;
-}
-
-.skill_container {
-  flex-wrap: wrap;
 }
 
 .infos {
